+++ scripts/report-issue.js
@@ -44,7 +44,7 @@ const collectSystemInfo = () => {
 		os: `${os.arch()}; ${os.version()}`,
 		nodeVersion: execSync("node -v").toString().trim(),
 		npmVersion: execSync("npm -v").toString().trim(),
-		clineVersion: getClineVersion(),
+		codaiVersion: getClineVersion(),
 	}
 }
 
@@ -66,7 +66,7 @@ const createIssueUrl = (systemInfo, issueTitle) => {
 		`https://github.com/cline/cline/issues/new?template=bug_report.yml` +
 		`&title=${issueTitle}` +
 		`&operating-system=${systemInfo.os}` +
-		`&cline-version=${systemInfo.clineVersion}` +
+		`&cline-version=${systemInfo.codaiVersion}` +
 		`&system-info=${
 			`Node: ${systemInfo.nodeVersion}\n` +
 			`npm: ${systemInfo.npmVersion}\n` +