+++ src/core/controller/task/methods.ts
@@ -6,8 +6,10 @@ import { registerMethod } from "./index"
 import { askResponse } from "./askResponse"
 import { cancelTask } from "./cancelTask"
 import { clearTask } from "./clearTask"
+import { deleteAllTaskHistory } from "./deleteAllTaskHistory"
 import { deleteNonFavoritedTasks } from "./deleteNonFavoritedTasks"
 import { deleteTasksWithIds } from "./deleteTasksWithIds"
+import { executeQuickWin } from "./executeQuickWin"
 import { exportTaskWithId } from "./exportTaskWithId"
 import { getTaskHistory } from "./getTaskHistory"
 import { getTotalTasksSize } from "./getTotalTasksSize"
@@ -23,8 +25,10 @@ export function registerAllMethods(): void {
 	registerMethod("askResponse", askResponse)
 	registerMethod("cancelTask", cancelTask)
 	registerMethod("clearTask", clearTask)
+	registerMethod("deleteAllTaskHistory", deleteAllTaskHistory)
 	registerMethod("deleteNonFavoritedTasks", deleteNonFavoritedTasks)
 	registerMethod("deleteTasksWithIds", deleteTasksWithIds)
+	registerMethod("executeQuickWin", executeQuickWin)
 	registerMethod("exportTaskWithId", exportTaskWithId)
 	registerMethod("getTaskHistory", getTaskHistory)
 	registerMethod("getTotalTasksSize", getTotalTasksSize)